Change data types to fix error

+++ OANDARestLibrary/TradeLibrary/DataTypes/Order.cs
@@ -14,17 +14,17 @@
 
         public double lowerBound { get; set; }
 
-        public double price { get; set; }
+        public decimal price { get; set; }
 
         public string side { get; set; }
 
-        public double stopLoss { get; set; }
+        public decimal stopLoss { get; set; }
 
-        public double takeProfit { get; set; }
+        public decimal takeProfit { get; set; }
 
         public string time { get; set; }
 
-        public int trailingStop { get; set; }
+        public decimal trailingStop { get; set; }
 
         public string type { get; set; }
